Decomp func_8005C708

+++ include/function_symbols.h
@@ -35,7 +35,7 @@ extern void func_8005C250(u16);
 extern void func_8005C520(s8, s8);
 extern s32 func_8005C550(u16 actor_index, s16 arg1);
 extern s32 func_8005C6D0(s32);
-extern s16 func_8005C708(s32);
+extern s32 func_8005C708(s32);
 extern s32 func_8005C774(s32);
 extern void func_8005CA34(s32 mag, s32 time); // shake camera vertically.
 extern s32 func_8005D1B0(u16);
